+++ source/app/graph/componentmanager.h
@@ -128,7 +128,7 @@ private slots:
 
 public:
     const std::vector<ComponentId>& componentIds() const;
-    int numComponents() const { return static_cast<int>(componentIds().size()); }
+    size_t numComponents() const { return componentIds().size(); }
     bool containsComponentId(ComponentId componentId) const;
     const GraphComponent* componentById(ComponentId componentId) const;
     ComponentId componentIdOfNode(NodeId nodeId) const;
